requirements.txt
click==8.1.7
colorama==0.4.6
construct==2.10.70
dearpygui==1.10.1
funcy==2.0
loguru==0.7.2
lxml==5.0.0
networkx==3.2.1
numpy==1.26.2
pillow==10.2.0
pydot==2.0.0
pylivemaker==1.1.2
pyparsing==3.1.1
win32-setctime==1.1.0
requests~=2.31.0